--- src/app/core/models/map.ts
@@ -1,3 +1,0 @@
-export interface Map<T> {
-  [key: string]: T;
-}
